Added Code for SQL injection prevention
Regex filtering would ensure there are no spaces or non-alphanumeric characters in the password provided, thus preventing the injection of queries or other forms of code. 

NOTE: Regex may be altered to allow certain special characters if deemed necessary, but for now we will look at alphanumeric (A-Z in caps and lower case and 0-9 digits) only.